Added feature to recognize, from which page was the updated initiated and subsequentially return to that page
explanation (a bit tricky)
This feature is quite tricky, since HTTP is stateless protocol, i had to manually implement in each view, where i am calling allPosts partial its url. That url got stored in each post edit form (GET) in form of hidden input (to be sent as query parameter). Subsequentially, the form method made GET req for post update page with the query, that got processed on the server and sent to form_update view additional argument with the name of redirectPathUpdate (from the query). Then in the form_update view i implemented another hidden input, in which the argument with url got stored and got sent to the server with POST req to update the post. The server processed req and if updated was succesfull, server redirected to the intial url